% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_control.R
\name{case_control}
\alias{case_control}
\title{case_control}
\usage{
case_control(p, d, dict, exposure_cds, visit_cd = "V", outcome_cd = "O",
  riskWindow_daysPreIndex_start = 365, riskWindow_daysPreIndex_end = 1,
  controls_num = 3)
}
\arguments{
\item{p}{A data frame of patients demographics}

\item{d}{A data frame of patient data in a tall file}

\item{dict}{A data frame of descriptions for each concept_cd in the data_tbl}

\item{exposure_cds}{A string vector of exposures codes to test}

\item{visit_cd}{The code used for visits (default is V)}

\item{outcome_cd}{The code used for the outcome in the data_tbl (default is O)}

\item{riskWindow_daysPreIndex_start}{The number of days prior to the index date to start the risk window (default is 365)}

\item{riskWindow_daysPreIndex_end}{The number of days prior to the index date to end the risk window (default is 1)}

\item{controls_num}{Number of controls to match to each case}
}
\value{
A list with the elements
\item{result}{A data frame with results including a row for each defined exposure}
\item{match_results}{A list returned from the matching function }
\item{cohort_file}{The full patient data frame with case control status after matching }
}
\description{
The main case control function to run a case control analysis from i2b2-generated data files
}
\examples{
#Not run
}
