% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ri2b2casecontrol.R
\docType{package}
\name{Ri2b2casecontrol}
\alias{Ri2b2casecontrol}
\alias{Ri2b2casecontrol-package}
\title{\code{Ri2b2casecontrol} package}
\description{
i2b2 R package to run case-control analysis
}
\details{
See the README on
\href{https://gitlab.partners.org/vc070/Ri2b2casecontrol}{GitHub}
}
